backend/requirements.lock.txt
asgiref==3.11.0
attrs==25.4.0
dj-rest-auth==7.0.2
Django==5.2.10
django-allauth==65.13.1
django-filter==25.2
djangorestframework==3.16.1
drf-spectacular==0.29.0
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==25.0
psycopg==3.3.2
psycopg-binary==3.3.2
psycopg2-binary==2.9.11
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
wheel==0.45.1
whitenoise==6.11.0
